refactor: move the tap test wire types to the tap types module
