+++ packages/app/src/tap/commands/test-state.ts
@@ -1,35 +1,6 @@
 import type { SerializedTest } from '@packages/types'
 
-import type { TapTestsRunner } from '../types'
-
-export type TestStateValue = 'passed' | 'failed' | 'pending' | 'skipped'
-
-export interface TestStateEntry {
-  id: string
-  title: string
-  duration?: number
-  state: TestStateValue
-  /** Retries actually taken this run, not the configured maximum. */
-  retries?: number
-}
-
-export interface TestError {
-  name?: string
-  message?: string
-  stack?: string
-}
-
-export interface TestDetailEntry {
-  id: string
-  title: string
-  /** Suite titles leading to this test plus its own, joined with ` > `. */
-  fullTitle: string
-  duration?: number
-  state: TestStateValue
-  retries?: number
-  timings?: Record<string, unknown>
-  error?: TestError
-}
+import type { TapTestsRunner, TestDetailEntry, TestError, TestStateEntry, TestStateValue } from '../types'
 
 // A state-less test was never reached: 'pending' while the run is still going,
 // 'skipped' once it is complete (matching the driver's end-of-run summary).
